Thanks to mroblivious1bmf I've implemented ANSI detection according to the *real* ANSI spec a public spec for terminals covering the same stuff can be found online at: http://www.ecma-international.org/publications/files/ecma-st/ECMA-048.pdf
SyncTERM now detects ANSI sequences according to the spec (the implemetation
still has holes) and so if you see the ESC[ thing now, it's due to a broken
ANSI code... SyncTERM isn't broken, the ANSI code is (honest).

In reading over the spec, it even shows a way to make both ANSI music and
delete line work most of the time (whee!).

Bump SyncTERM version to 0.5